Don't play wrapping games
The driver calls the wrapped CloseScreen function in the middle of messing
with the hardware state.  On multi-card setups this may cause the VGA
arbiter to switch to a different card while we're in the middle of things,
with obvious disastrous effects.  This fixes things by making sure we
only call the wrapped CloseScreen function at the very end of the driver's
CloseScreen function.

Signed-off-by: Mark Kettenis <kettenis@openbsd.org>