Don't gratuitously disable Render when ShadowFBing
There's no reason to do this, and I am at a loss to explain why that was
ever done.  The first instance of this logic comes from xfree86 (over 11
years ago!):

http://cvsweb.xfree86.org/cvsweb/xc/programs/Xserver/hw/xfree86/drivers/ati/atiscreen.c.diff?r1=1.15&r2=1.16

Now, cfb never had Render support, so that might make some sense if
there was an intermediate uncommitted state where the driver used fb for
direct but cfb for shadowed.  But that's really the only plausible
explanation I can think of.

Signed-off-by: Adam Jackson <ajax@redhat.com>