Remove dummy UTS / DFS hooks
When EXA was added to mach64, it included UploadToScreen and
DownloadFromScreen hooks that did the same thing that EXA would do in a
software fallback. Now they are out of date and cause crashes. The
slight benefit of reimplementing those hooks (if there is any) is not
worth the maintenance burden of having to rewrite them every time there
is a change to the way EXA handles pixmaps.
Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=51137

Signed-off-by: Connor Behan <connor.behan@gmail.com>
Tested-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>