Remove unused atiaudio.c & ATIAudioNames declaration
Defined a string array, which is not used anywhere.

atiaudio.h remains for now, though for the enum it defines, only one
reference to it remains, initializing the Audio to ATI_AUDIO_NONE, so
it may be a candidate for further cleanup in the future.

Reported by Solaris linker -z guidance option:
ld: guidance: removal of unused file recommended: atiaudio.o

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>