Fix compositing rotation
This corrects the composite operation's interpretation of a
source picture transformation matrix indicating rotation,
correcting a previous behavior where the source image was simply
flipped and not rotated. This is done by using a transformed vector
for each vertex of the source rectangle, instead of just two at
each of the diagonally opposed corner vertices.